refactored to implement .env and other changes
