+++ src/main/java/com/nukkadseva/nukkadsevabackend/service/implementation/ProviderServiceImpl.java
@@ -1,6 +1,6 @@
 package com.nukkadseva.nukkadsevabackend.service.implementation;
 
-import com.nukkadseva.nukkadsevabackend.dto.ProviderDto;
+import com.nukkadseva.nukkadsevabackend.dto.request.ProviderDto;
 import com.nukkadseva.nukkadsevabackend.dto.response.ProviderSummaryDto;
 import com.nukkadseva.nukkadsevabackend.entity.Provider;
 import com.nukkadseva.nukkadsevabackend.entity.Users;
